refactor : rename and move some functions around
