+++ src/class-decor/mixer.ts
@@ -1,8 +1,7 @@
 import {Class, Instance} from "../core/types";
 import {classPrivateState, ClassStateProvider} from "../core/class-private-state";
 import {initEdgeClass} from "./apply-method-decorations";
-import {AfterHook, BeforeHook, MiddlewareHook} from "../functoin-decor/index";
-import {FunctionMetaData} from "../functoin-decor/wrapper";
+import {AfterHook, BeforeHook, MiddlewareHook, FunctionMetaData} from "../functoin-decor/index";
 
 type DumbClass = new(...args: any[]) => object;
